splitted code into modules and added toString func
