+++ colorcoder/ColorNames.java
@@ -1,0 +1,12 @@
+package colorcoder;
+
+public class ColorNames {
+    final static String MajorColorNames[] = {
+        "White", "Red", "Black", "Yellow", "Violet"
+    };
+    final static int numberOfMajorColors = MajorColorNames.length;
+    final static String MinorColorNames[] = {
+        "Blue", "Orange", "Green", "Brown", "Slate"
+    };
+    final static int numberOfMinorColors = MinorColorNames.length;
+}
