+++ colorcoder/Getters.java
@@ -1,0 +1,15 @@
+package colorcoder;
+
+public class Getters {
+    static ColorPair GetColorFromPairNumber(int pairNumber) {
+        int zeroBasedPairNumber = pairNumber - 1;
+        MajorColor majorColor = 
+            MajorColor.fromIndex(zeroBasedPairNumber / ColorNames.numberOfMinorColors);
+        MinorColor minorColor =
+            MinorColor.fromIndex(zeroBasedPairNumber % ColorNames.numberOfMinorColors);
+        return new ColorPair(majorColor, minorColor);
+    }
+    static int GetPairNumberFromColor(MajorColor major, MinorColor minor) {
+        return major.getIndex() * ColorNames.numberOfMinorColors + minor.getIndex() + 1;
+    }
+}
