+++ colorcoder/Tests.java
@@ -1,0 +1,23 @@
+package colorcoder;
+
+public class Tests {
+    static void testNumberToPair(int pairNumber,
+        MajorColor expectedMajor,
+        MinorColor expectedMinor)
+    {
+        ColorPair colorPair = Getters.GetColorFromPairNumber(pairNumber);
+        System.out.println("Got pair " + colorPair.ToString());
+        assert(colorPair.getMajor() == expectedMajor);
+        assert(colorPair.getMinor() == expectedMinor);
+    }
+
+    static void testPairToNumber(
+        MajorColor major,
+        MinorColor minor,
+        int expectedPairNumber)
+    {
+        int pairNumber = Getters.GetPairNumberFromColor(major, minor);
+        System.out.println("Got pair number " + pairNumber);
+        assert(pairNumber == expectedPairNumber);
+    }
+}
